Added new features --parallel-request.
With the new cmdline opiton --parallel-request, all targets given on the command line will be processed in parallel.

This is implemented by using an additional transparent "phony:@parallel-request" target. This way, this feature does not require any change to the target executor. The downside is an additional systhetic target which might confuse the user a little bit. Also, the different colors for requested targets vs. dependencies is not correct atm.