Reworked classloaders parallel capability.
According to the JavaDoc of java.lang.ClassLoader, classloaders that are not strictly hierachical, which is the case for ProjectClassLoader and PluginClassLoader, need to be parallel capable, to avoid deadlocks. Now, these deadlocks might be the ones we tried to hunt down by every tick in the book, but without the essential clue. I need now to run further tests but all known scenarios we used to reproduce, now work. :-)