Fixed broken parallel classloader registration.
To register a classloader as parallel capable, the protected static caller sensitive method ClassLoader.registerAsParallelCapable() must be called.
As this is not possible in a static way with Scala, we create a dummy instance call the method from there then drop that dummy instance and only after that, start to use newly created instances of that classloader. We do this for the ProjectClassLoader and the PluginClassLoader.